Adding Option -bi to process adb shell dumpsys batteryinfo output
This allows processing of this lone output, rather than having to
process an enentire bugreport, or to insert the ==== dumpstate: =====
header in order to to skip pass sanity test and act like full
bugreport.  Previously, processing of batteryinfo alone was simply
treated as a partial header, and no additional csv files etc. were ever
output.