Adding --sqlite option to silence all output except sqlite.db creation
Added --sqlite option to output the log and the report.db sqlite
database only.  Basically, just wrapped csv, png and html chapter
printout in an if statement that is dependent on a --sqlite option.
Option is stored as boolean in Context, similar to processing of
--silent.  Some of these conditions may be redundant given the return
from all chapter rendering in doc/Chapter.java… so can probably delete
many of these...